Use implementation from RustCrypto ed25519
