+++ Cargo.toml
@@ -18,6 +18,7 @@ bitflags = "1.2.1"
 byteorder = "1.4.2"
 chrono = "0.4.19"
 crc16 = "0.4.0"
+ed25519 = "1.3.0"
 json = "0.12.0"
 num-traits = "0.2.12"
 num-bigint = "0.2.0"
@@ -25,7 +26,20 @@ num-rational = "0.3.0"
 rust_decimal = "1.6.0"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
-sodiumoxide = "0.2.5"
+sha2 = "0.10.1"
 thiserror = "1.0.20"
 xdr-rs-serialize = "0.2.4"
 xdr-rs-serialize-derive = "0.2.4"
+
+[dependencies.sodiumoxide]
+version = "0.2.5"
+optional = true
+
+[dev-dependencies]
+ed25519-dalek = "1.0.1"
+rand = "0.7.3"
+ring-compat = "0.4.0"
+
+[features]
+default = ["sodium_oxide"]
+sodium_oxide = ["sodiumoxide"]
