+++ src/crypto/mod.rs
@@ -1,42 +1,85 @@
 //! Cryptographic functions.
-//!
-//! For long running programs, you should initialize the sodium
-//! library for optimal performance.
-//!
-//! ```rust
-//! use stellar_base::crypto;
-//!
-//! # fn run() -> stellar_base::error::Result<()> {
-//! crypto::init()?;
-//! // your program here.
-//! # Ok(())
-//! # }
-//! ```
-use crate::error::{Error, Result};
-use sodiumoxide::crypto::hash::sha256;
-use sodiumoxide::randombytes;
 
-mod keypair;
+mod public_key;
+mod signature;
+#[cfg(feature = "sodium_oxide")]
+mod sodium_oxide;
 mod strkey;
 
-pub use self::keypair::{KeyPair, MuxedAccount, MuxedEd25519PublicKey, PublicKey, SecretKey};
+use sha2::Digest;
+
+pub use self::public_key::{MuxedAccount, MuxedEd25519PublicKey, PublicKey};
+pub use self::signature::*;
 pub use self::strkey::*;
+pub use ed25519::signature::{Signer as Ed25519Signer, Verifier as Ed25519Verifier};
+pub use ed25519::Signature;
+#[cfg(feature = "sodium_oxide")]
+pub use sodium_oxide::*;
 
 /// Compute sha256 hash of `m`.
 pub fn hash(m: &[u8]) -> Vec<u8> {
-    let digest = sha256::hash(&m);
-    digest.0.to_vec()
+    sha2::Sha256::digest(&m).to_vec()
 }
 
-/// Generate `size` random bytes.
-pub fn random_bytes(size: usize) -> Vec<u8> {
-    randombytes::randombytes(size)
+#[derive(Debug, Clone, PartialEq)]
+pub struct SigningKey<S>
+where
+    S: Ed25519Signer<Signature>,
+{
+    pub signing_key: S,
 }
 
-/// Initialize the sodium library and chooses faster version of the primitives
-/// if possible.
-///
-/// `init` also makes `KeyPair::random()` thread-safe.
-pub fn init() -> Result<()> {
-    sodiumoxide::init().map_err(|_| Error::SodiumInitFailed)
+#[derive(Debug, Clone, PartialEq)]
+pub struct VerifyKey<V>
+where
+    V: Ed25519Verifier<Signature> + AsRef<[u8]>,
+{
+    pub verify_key: V,
+}
+
+#[derive(Debug, Clone, PartialEq)]
+pub struct KeyPair<S, V>
+where
+    S: Ed25519Signer<Signature>,
+    V: Ed25519Verifier<Signature> + AsRef<[u8]>,
+{
+    pub signer: SigningKey<S>,
+    pub verifier: VerifyKey<V>,
+}
+
+impl<S, V> KeyPair<S, V>
+where
+    S: Ed25519Signer<Signature>,
+    V: Ed25519Verifier<Signature> + AsRef<[u8]>,
+{
+    pub fn new(signer: S, verifier: V) -> KeyPair<S, V> {
+        KeyPair {
+            signer: SigningKey {
+                signing_key: signer,
+            },
+            verifier: VerifyKey {
+                verify_key: verifier,
+            },
+        }
+    }
+
+    pub fn sign(&self, msg: &[u8]) -> Signature {
+        self.signer.signing_key.sign(msg)
+    }
+
+    pub fn verify(&self, msg: &[u8], sig: &Signature) -> Result<(), ed25519::Error> {
+        self.verifier.verify_key.verify(msg, sig)
+    }
+
+    /// Sign the `message` together with the signature hint.
+    pub fn sign_decorated(&self, message: &[u8]) -> DecoratedSignature {
+        let hint = self.signature_hint();
+        let signature = self.signer.signing_key.sign(message);
+        DecoratedSignature::new(hint, signature)
+    }
+
+    /// Return the signature hint, that is the last 4 bytes of the public key.
+    pub fn signature_hint(&self) -> SignatureHint {
+        SignatureHint::from_public_key(&self.verifier.verify_key.as_ref())
+    }
 }
