+++ src/lib.rs
@@ -16,10 +16,10 @@
 //! to sign transactions.
 //!
 //! ```rust
-//! use stellar_base::crypto::KeyPair;
+//! use stellar_base::crypto::{SodiumKeyPair};
 //!
 //! # fn run() -> stellar_base::error::Result<()> {
-//! let random_kp = KeyPair::random()?;
+//! let random_kp = SodiumKeyPair::random()?;
 //! println!("Account Id = {}", random_kp.public_key().account_id());
 //! # Ok(())
 //! # }
@@ -34,7 +34,7 @@
 //! ```rust
 //! use stellar_base::amount::Amount;
 //! use stellar_base::asset::Asset;
-//! use stellar_base::crypto::{KeyPair, PublicKey};
+//! use stellar_base::crypto::{PublicKey, SodiumKeyPair};
 //! use stellar_base::memo::Memo;
 //! use stellar_base::network::Network;
 //! use stellar_base::operations::Operation;
@@ -43,7 +43,7 @@
 //! use std::str::FromStr;
 //!
 //! # fn run() -> stellar_base::error::Result<()> {
-//! let source_kp = KeyPair::random()?;
+//! let source_kp = SodiumKeyPair::random()?;
 //! let destination = PublicKey::from_account_id("GATTMQEODSDX45WZK2JFIYETXWYCU5GRJ5I3Z7P2UDYD6YFVONDM4CX4")?;
 //!
 //! let payment_amount = Amount::from_str("13.12")?;
@@ -59,7 +59,7 @@
 //!     .add_operation(payment)
 //!     .into_transaction()?;
 //!
-//! tx.sign(&source_kp, &Network::new_test());
+//! tx.sign(&source_kp.as_ref(), &Network::new_test());
 //! let xdr = tx.into_envelope().xdr_base64()?;
 //! println!("Xdr = {}", xdr);
 //! # Ok(())
@@ -85,7 +85,6 @@ pub mod memo;
 pub mod network;
 pub mod operation_result;
 pub mod operations;
-pub mod signature;
 pub mod time_bounds;
 pub mod transaction;
 pub mod transaction_result;
@@ -93,7 +92,7 @@ pub mod xdr;
 mod xdr_generated;
 
 pub use self::asset::Asset;
-pub use self::crypto::{KeyPair, PublicKey};
+pub use self::crypto::PublicKey;
 pub use self::memo::Memo;
 pub use self::network::Network;
 pub use self::operation_result::OperationResult;
