+++ src/operations/end_sponsoring_future_reserves.rs
@@ -62,23 +62,12 @@ impl EndSponsoringFutureReservesOperationBuilder {
 
 #[cfg(test)]
 mod tests {
-    use crate::crypto::KeyPair;
+    
     use crate::network::Network;
+    use crate::operations::tests::*;
     use crate::operations::Operation;
     use crate::transaction::{Transaction, TransactionEnvelope, MIN_BASE_FEE};
     use crate::xdr::{XDRDeserialize, XDRSerialize};
-
-    fn keypair0() -> KeyPair {
-        // GDQNY3PBOJOKYZSRMK2S7LHHGWZIUISD4QORETLMXEWXBI7KFZZMKTL3
-        KeyPair::from_secret_seed("SBPQUZ6G4FZNWFHKUWC5BEYWF6R52E3SEP7R3GWYSM2XTKGF5LNTWW4R")
-            .unwrap()
-    }
-
-    fn keypair1() -> KeyPair {
-        // GAS4V4O2B7DW5T7IQRPEEVCRXMDZESKISR7DVIGKZQYYV3OSQ5SH5LVP
-        KeyPair::from_secret_seed("SBMSVD4KKELKGZXHBUQTIROWUAPQASDX7KEJITARP4VMZ6KLUHOGPTYW")
-            .unwrap()
-    }
 
     #[test]
     fn test_end_sponsoring_future_reserves() {
@@ -87,7 +76,7 @@ mod tests {
             .add_operation(Operation::new_end_sponsoring_future_reserves().build())
             .into_transaction()
             .unwrap();
-        tx.sign(&kp, &Network::new_test()).unwrap();
+        tx.sign(&kp.as_ref(), &Network::new_test()).unwrap();
         let envelope = tx.to_envelope();
         let xdr = envelope.xdr_base64().unwrap();
         let expected = "AAAAAgAAAADg3G3hclysZlFitS+s5zWyiiJD5B0STWy5LXCj6i5yxQAAAGQADKI/AAAAAwAAAAAAAAAAAAAAAQAAAAAAAAARAAAAAAAAAAHqLnLFAAAAQOQlXWOVDQVXDDwpbjv6EpROZHQ77tzmTMCrczR20UFqzPBZHc6lKv9tTLzz/esVDbwYbaa+m/y8HqH7MqPjuQw=";
@@ -109,7 +98,7 @@ mod tests {
             )
             .into_transaction()
             .unwrap();
-        tx.sign(&kp, &Network::new_test()).unwrap();
+        tx.sign(&kp.as_ref(), &Network::new_test()).unwrap();
         let envelope = tx.to_envelope();
         let xdr = envelope.xdr_base64().unwrap();
         let expected = "AAAAAgAAAADg3G3hclysZlFitS+s5zWyiiJD5B0STWy5LXCj6i5yxQAAAGQADKI/AAAAAwAAAAAAAAAAAAAAAQAAAAEAAAAAJcrx2g/Hbs/ohF5CVFG7B5JJSJR+OqDKzDGK7dKHZH4AAAARAAAAAAAAAAHqLnLFAAAAQBjITBRZCRbpp2adVQSTC8l9NB/UoTFtHpHkZYEEWl0hSGD7U4N3m2WhFD+kmoBnyy1jWbtPMkyL3SriGvrSCwM=";
