+++ src/operations/tests.rs
@@ -1,0 +1,19 @@
+use crate::crypto::SodiumKeyPair;
+
+pub fn keypair0() -> SodiumKeyPair {
+    // GDQNY3PBOJOKYZSRMK2S7LHHGWZIUISD4QORETLMXEWXBI7KFZZMKTL3
+    SodiumKeyPair::from_secret_seed("SBPQUZ6G4FZNWFHKUWC5BEYWF6R52E3SEP7R3GWYSM2XTKGF5LNTWW4R")
+        .unwrap()
+}
+
+pub fn keypair1() -> SodiumKeyPair {
+    // GAS4V4O2B7DW5T7IQRPEEVCRXMDZESKISR7DVIGKZQYYV3OSQ5SH5LVP
+    SodiumKeyPair::from_secret_seed("SBMSVD4KKELKGZXHBUQTIROWUAPQASDX7KEJITARP4VMZ6KLUHOGPTYW")
+        .unwrap()
+}
+
+pub fn keypair2() -> SodiumKeyPair {
+    // GB7BDSZU2Y27LYNLALKKALB52WS2IZWYBDGY6EQBLEED3TJOCVMZRH7H
+    SodiumKeyPair::from_secret_seed("SBZVMB74Z76QZ3ZOY7UTDFYKMEGKW5XFJEB6PFKBF4UYSSWHG4EDH7PY")
+        .unwrap()
+}
